Update playTime for moveOne()

+++ arduino-motor-drivers/lib/HerkulexMotor/HerkulexMotor.h
@@ -47,8 +47,8 @@ class HerkulexMotor{
     private:
         int _id;
         MotorModel _type;
-        int _zeroPos;
-        int _bounds[2];
+        uint16_t _zeroPos;
+        uint16_t _bounds[2];
 };
 
 #endif
